minor fixes

+++ routes/book.routes.js
@@ -36,10 +36,10 @@ router.put("/update-book", authenticateToken,async(req,res)=>{
         //to check user is admin or not
         const {bookid} = req.headers;
         
-        // if(user.role !== "admin"){
-        //     return res.status(403).json({ message: "You are not an admin" });
-        // }
-        //book
+        if(user.role !== "admin"){
+            return res.status(403).json({ message: "You are not an admin" });
+        }
+        book
         const { url,title, author, price,description,language } = req.body;
         await Book.findByIdAndUpdate(bookid,{
             url: url,
@@ -62,10 +62,10 @@ router.delete("/delete-book", authenticateToken,async(req,res)=>{
         //to check user is admin or not
         const {bookid} = req.headers;
         
-        // if(user.role !== "admin"){
-        //     return res.status(403).json({ message: "You are not an admin" });
-        // }
-        //book
+        if(user.role !== "admin"){
+            return res.status(403).json({ message: "You are not an admin" });
+        }
+        book
         const { url,title, author, price,description,language } = req.body;
         await Book.findByIdAndDelete(bookid,{
             url: url,
